rect: Document the type semantics
The graphene_rect_t allows flipping the origin using negative sizes, but
we always expect the public entry points to normalize it so that we can
assume positive sizes when operating on them.

The public API does not act like a mutator: we always implicitly operate
on a normalized copy internally.