tests/vec4: Use a larger epsilon
We end up hitting fluctuations in the floating point representation, so
it's better if we use a larger epsilon to check for the component of a
vec4 after normalization.

This should probably be applied to the other vector test units as well.

Tested with the SSE2, GCC vector fast paths, as well as the scalar
implementation.