Add the prototype of a benchmark suite
We want to test the various implementations of the fast paths inside
Graphene, so we should have at least some form of benchmarking suite
for them.

The suite is private, but its API is modelled on GTest and the GObject
performance tests. It's not perfect: the build is especially clunky and
needs further refactoring for the rules, so that we can avoid
replicating them all just because we change a pre-processor define.