Rework the implementation of the SIMD types
The current implementation of the SIMD types is slightly less useful
than it ought to be, and we should make it better. The main goal of this
commit is to make the SIMD vector and matrix API public and usable
outside of Graphene itself, through the judicious application of
inlining and macros.

We can inline everything inside the simd4f.h and simd4x4f.h headers, and
use compiler extensions to macroize all the platform-specific API using
a trick similar to the one used by gatomic.[ch] in GLib. Currently, this
means that we support only GCC and GCC-compatible compilers, like Clang;
supporting SunCC and MSVC would be nice, but it'll require some help
from people that have access to those compiler suites.