Add custom allocator for SIMD-backed types
While we're guaranteed the alignment of SIMD types on the stack, the
same guarantees do not apply when allocating them on the heap. We can
assume that malloc() will align to multiples of 8 and 16 on basically
all the platforms we care about, but it's not something we can take for
granted.

Luckily, we can check if the platform we're compiling for has an
allocator function that supports alignment; with the advent of ISO C11
and its aligned_alloc() the chances that the platform has that function
or one close to it in semantics and behaviour, the check is likely to
turn out something we can use.