Release 0.99.2 - release candidate 1
Now that the documentation hit the 100% mark, and we can generate
introspection data, I think it's time for a Graphene 1.0 release.

The "1.0" moniker does not imply that the code is Bug Free™: I still
haven't tested all the math, and the test suite coverage is painfully
low. Plus, the introspection generation depends on an unreleased
version of GObject-Introspection, so there's that.

The API, on the other hand, is pretty much what I want it to look like.
I think that there are still some minor adjustments to be made in some
of the modifiers, like the vector length and normalization functions
modifying the vector in place, instead of returning a modified copy; on
the other hand, I feel pretty confident that there won't be major
changes at this point, and that any other API work on Graphene is going
to rely on additions and deprecations.

The other reason why I want to do a 1.0 release is that I'll soon start
publishing the code I'm working on that depends on the Graphene API,
which means that I'll need an actual release.