build: Look for NEON intrinsics instead of assembly
We're checking for availability of ARM NEON assembly in order to
determine whether or not we can use NEON instructions; at the same time,
we're not using assembly anywhere, and we're instead relying on the
existence of the NEON intrinsics. There's the chance of false positives
out there — toolchains that only have ARM NEON assembly but no
intrinsics supported — so we need to be more explicit as to what we look
for and use. If, at some point, we'll have assembly versions of the NEON
SIMD implementation then we'll add the assembly check back.