Add a plane type
Planes are useful in fully 3D canvases, as they can be used for things
like frustum-based culling.

The graphene_plane_t data structure represents a plane via a normal
vector pointing towards the origin of the plane; and a constant value
representing the distance from the origin along the normal vector.