graphene-simd4x4f.h: Allow Build On Visual Studio
Fix the initialization in graphene_simd4x4f_init() so that compilers will
be happy with the types, and add a definition for
graphene_simd4x4f_transpose_in_place(s) that is suitable for Visual Studio
when we have SSE2 intrinistic support.  The SSE2 initrinistic support, due
to ABI issues, like the GCC builds, are only supported on x64 builds, and
possibly x86 builds on MSVC 2013 and later using __vectorcall.