box: Use vec3 instead of point3d
We use vec3 math internally, and converting back and forth between
coordinates and vectors is not nearly as nice as converting only at
specific entry points.

This requires some surgery for the static array of constant boxes, but
we can follow the same pattern used for the static array of constant
vectors we already have.