build: Add mapfile for Solaris builds
I originally nibbed this whole chunk from the Pixman configure.ac, but I
forgot to modify it and include the mapfile. Since nobody complained, it
is clear that Graphene hasn't been built on Solaris.

I do not have access to a Solaris installation, nor to SunCC, so I'm
tempted to just bin the whole thing and wait for outside contibutions;
since it does not cost me anything, yet, I'll just leave it here for a
while. I consider SunCC/Solaris support on notice, though, and will
likely remove it if nothing happens after I release 1.2.