rect: Rename an internal function
Rename normalize_internal to normalize_in_place.

It does not make *much* of a difference, since the public entry point
does exactly the same, with a return value, but it makes it clear that
the rectangle is normalized in place.

This is a pretty bad API wart; the public entry point should take an
additional argument, like we do everywhere else.