docs: Include a coding style document
Since we are getting contributions, and I don't want to spend time
explaining coding style in the future, let's be explicit about the
kind of style I'm going to accept.

This is just the GLib/GTK+/Clutter coding style, amended where it
makes more sense in the context of Graphene.