euler: Fix the math for the matrix conversion
We are using a left-handed sign convention, to ensure that creating a
matrix through graphene_euler_t, graphene_quaternion_t, and using the
angle-axis notation results in the same rotation matrix.

We need to document the exact nature of our transformation, including
the matrices used to compute the resulting row vectors.