travis: Update the GCC toolchain
Sadly, Travis CI is using GCC 4.6 from the Ubuntu 12.04 LTS
repositories, which is apparently buggy, especially when it comes to
floating point math optimizations. Since floating point math is all that
Graphene does, bumping up the toolchain to use a slightly more recent
(and correct) GCC suite seems the only way to ensure that the test suite
passes on Travis with GCC just like it already does on Travis with Clang
and locally, with GCC 4.9.