vectors: Allow fuzzyness in equality checks
We should allow for floating point representation discrepancies when
checking two vectors for equality. The immediate check using the SIMD
comparison operator has the same issue as a direct equality using
memcmp() or '==', but also we want to check within an arbitrary high
precision.

We use the same strategy (and epsilon) to check if two graphene_point_t
or two graphene_point3d_t are equal as well.