matrix: Clarify the lerp behaviour with non-decomposable matrices
Right now we leave the behaviour of interpolating two matrices that
cannot be decomposed into their transformations unspecified; we keep the
result matrix as it is, by applying 'null' transformations, but the
matrix that is passed in for the result could contain anything — it
could even be uninitialized, and the caller could rely on Graphene
putting sane values inside it.

We should provide a stable behaviour for our API, and document it
properly. The sane approach is to just bail out early, and initialize to
an identity matrix. We don't do that unconditionally to avoid
unnecessary cost, since, in case of success, the identity matrix will be
blasted away in any case.