Ensure that we don't introspect system headers
The introspection scanner has issues parsing SIMD system headers and
compiler-specific extensions. This creates a bunch of warnings that we
can't do anything about without a full overhaul of the introspection
scanner.

Luckily, we can detect when a header file is being included via the
introspection scanner, and we can fall back to dummy scalar definitions.
After all, the scanner is only using the definitions for sizing
structures and for resolving names; the SIMD API is not accessible via
introspection.