build: Use an external script for filtering type names
When building introspection data we need to turn our lowercase type
names ending with '_t' into camel-case identifiers that match the
expectations of GObject-Introspection.

We currently use a Sed invocation, but we're also relying on GNU Sed
extensions that are not portable to other operating systems — e.g.
MacOS and *BSD.

Instead of using Sed, we can use a small Python script that does the
same transformation, with a higher chance at being portable.

Fixes #42