matrix: Go back to the simd4x4 is_2d() operator
Now that we have a working baseline, and now that the is_2d() operator
for graphene_simd4x4f_t is not affected by floating point fluctuations,
we can go back to using it.

We can leave the fuzzy comparison code in place, in case of regressions,
for ease of debugging.