rect: Add non-mutating versions of various methods
The graphene_rect_t API was written before I settled down on the design
tenet of providing functions that do not mutate the object.

The normalize(), inset(), and offset() methods operate in place on the
given rectangle; we should provide equivalent methods that do not mutate
the rectangle they are given but, instead, store the result in a
separate out argument.