rect: Provide a replacement for round_to_pixel()
The round_to_pixel() method of graphene_rect_t is all kinds of
problematic.

First of all, it operates in place, whereas we want our mutators to
operate on a separate object; then, it mentions "pixels" in the name,
when it really has nothing to do with them; finally, it's not clear what
it does, as it's not similar to any other operator we have inside the
rectangle API.

We should simply have a round() method that operates on a copy of the
given rectangle, and is clearly tied to the concept of 'rounding' the
values of the rectangle's description.