docs: Clarify the modelview required when unprojecting
When unprojecting a point, Graphene expects the inverse of the modelview
used to project it.

We cannot really change the implementation to invert the modelview
ourselves without breaking existing code. There's also the common
behaviour of canvas libraries to cache the inverse of the modelview to
avoid incurring into additional costs, which means we'd be working
against our callers.

Fixes issue #50