matrix: Fix interpolation between 2D matrices
The decomposition and interpolation code claims to come from the CSS3
Transformation specification, but it does something fairly custom; it
also uses the existing 3D matrix interpolation and recomposition code
which is just not really correct.

We should split the interpolation cases for 2D and 3D matrices in a more
radical way, and follow the whole CSS3 Transformation spec more closely.