build: Clean up the Meson build
Depend on Meson ≥ 0.34.0 so we can use:

 * join()
 * is_odd()

This allows us to have an 'auto' setting for the `enable-debug`
configuration option, which enables debugging depending on whether we're
in a stable or unstable development cycle. It also allows us to simplify
the SIMD configuration variable by avoiding an unnecessary foreach.

We also can clean up some unused variables here and there.