build: Ensure soname versioning matches
We need to ensure that the soname used when installing the Graphene
shared object is the same when building with Autotools and when building
with Meson.

The current scheme, mutuated from GTK+, is:

  libgraphene-1.0.so
  `- .0
     `- (minor * 100 + micro - interface)
        `- interface

Where `interface` is equal to `micro`, unless it's reset to 0 — which
happens every time we add new API, during development cycles.

This means that, for version `1.4.1`, the soname of the installed shared
library is `0.400.1`, whereas for version `1.5.8`, the soname would be
`0.508.0`.