meson.build: Make it more MSVC-friendly
We are checking for compiler flags for warnings for GCC, but not Visual
Studio, so do a similar check based on those that we have in GLib's
msvc_recommended_pragmas.h when we are compiling for Visual Studio (the
advantage of being the one that maintains the MSVC builds files for the
GTK+ stack :)).  As a result, we avoid using GCC-style compiler/linker
flags, so that we don't choke meson's configure process and avoid
needless warnings during the build.

We also want to include SSE2 support for all x64 Visual Studio builds and
x86/32-bit Visual Studio 2013+ builds, as the SSE2 intrinsics are
supported on Visual Studio for a good while (since the 2008 days or
before), but as we will be passing in 4 __m128i's when we do
graphene_simd4x4f_init(), we can only support Visual Studio 2013 and later
as a special flag which is now in the code is supported to allow the code
to build with the 32-bit ABI.