build: Generate the API reference through Meson
The current build rules for the API reference are insufficient and
generally broken.

In order to fix them, we need to match what gtk-doc.make does, by
including a `gtkdocentitities.ent` template that is going to be included
by the main XML.

Additionally, we need to include the assets and fix the generation of
the ancillary files.

Currently, Meson will ignore `graphene-overrides.txt`, but the next
version will automatically copy it from srcdir to builddir.