Drop Travis CI
The Travis CI build environment is always hopelessly outdated, and we're
depending on various new components:

 * recent GCC and Clang
 * Ninja
 * Python 3
 * Meson

One way to handle CI for Graphene is to use Docker and build our own
custom image, but that requires some more work.

For the time being, we're going to rely on GNOME Continuous to actually
check that Graphene builds.