Update initialization macros for C99
Since Graphene requires a C99-compliant compiler to build, we should
also require a C99-compliant compiler to use it.

The point, point3d, size, and rect initialization macros should use
designated initializers, to ensure we're initializing the right  fields
while keeping the code legible.

Additionally, by using the type cast notation, we support using the
initialization macros as part of a compond literal, so we can do:

  graphene_matrix_translate (&m, &GRAPHENE_POINT3D_INIT (x, y, z));

and avoid a temporary variable.