src/: Fix build on Visual Studio (and possibly other compilers)
On Visual Studio at least, in static variables and global constants,
one is not allowed to have non-constant initializers[1], which will cause
the build to break with C2099 (initializer not constant).  So, use only
constants to initialize the global static constants that we now switched
to C99-style initialation macros, which unfortunately means that we have
to initialize them in the formoer way.

Note that these initializer macros work fine on Visual Studio otherwise,
including building GTK+-3.89.1.

[1]:
http://stackoverflow.com/questions/6131455/compile-error-c2099-initializer-is-not-a-constant,
please see first reply.