Fix build on older GCC versions
The fix for MSVC in commit 05b13b9972ade43115eba3204e45edbb2047aeb0 also
applies to older versions of GCC — but not of Clang.

Since there's really no reason to initialize static storage to zero, as
it's guaranteed by the C standard, we can simply drop the explicit
initialization altogether.