Update build docs for MSYS2
This commit updates README.py with better tested documentation for the
MSYS2 platform on Windows. Previously the instructions would fail because
"automake" is not a package.

This commit also gets the user to select packages based on their intended
target platform, a concept that matters on MSYS2 because each target
has its own $PREFIX.

The docs were also out of date in that they did not cover meson or the
Python dependency.