Use the appropriate Python when generating introspection data
We currently hardcode 'python' as the interpreter, but what we really
want is to reuse the same Python interpreter as the one we use for
Meson.

Luckily, Meson has a 'python3' module that we can use to find the
interpreter's path, and use that.

This allows us to drop an implicit dependency on Python 2.x.