Clean up the build rules
We should let Meson generate the pkg-config file for us, instead of
shipping a template file. We can rely on newer versions of Meson having
the right features to fill out the various variables, and add new ones
as well.

This has the effect that we can drop the global configuration_data
object and use smaller objects instead, one for each file we generate.