Drop the config.h template
The config.h template has the downside of going out of sync with the
configuration data we generate inside our build, as you need to modify
both the meson.build file, and the template itself.

Meson allows generating the config.h header without a template, using
the keys and values inside the configuration data.

This requires moving a bunch of keys and values to their own
configuration data object, instead of having a big bag of settings.