Release Graphene 1.8.0
No new API, this cycle, but it's time to spin a new stable release.

The major news for this release is that we dropped the Autotools build
from the repository; Graphene is now built exclusively with Meson.

We've added various fixes to the build on macOS, in order to maintain
compatibility with the old Autotools build.

The variables in the pkg-config file that expose the SIMD functionality
built inside Graphene have been updated, to ensure that they can be
easily consumed by other projects; instead of a single string that needs
to be split, we now expose boolean variables for all SIMD
implementations.

We've also improved the ability to use Graphene as a Meson sub-project;
this allows Meson to check out, build, and depend on Graphene when
building complex projects, like GTK. This is really useful inside CI
pipelines, as it minimises the amount of separate steps needed when
setting up the build.