build: don't hide the stdbool include from g-ir-scanner
g-i recently lost the ability to infer the bool type without a stdbool include.
While this is about to be fixed in g-i (https://gitlab.gnome.org/GNOME/gobject-introspection/merge_requests/116)
I don't see a reason why we should hide this in the first place, so remove the ifdef guard.