Use µTest for the Graphene test suite
GLib's GTest API is not that good, and it adds a dependency for basic
functionality like testing our API on multiple platforms.

The µTest library is smaller than the whole of GLib; it has a nicer API
and output than GTest; and we can use it as a Meson sub-project if it's
not installed in the system, which makes it very convenient to use.