rect: Docstring fix for rounded as it could not contain the original
A rectangle resulted to a call to `graphene_rect_round` could actually not
fully contain the original rectangle, in the sense that calling
`graphene_rect_contains` might fail when the difference between the floored
coordinates and the original one causes a shift.

So mention in the docstring that we guarantee that only the rounded area is
major or equal to the original one.

Also assert this in the test for rounding.