Add code of conduct
Graphene is used in various projects that are hosted on infrastructure
with a code of conduct, and it would be bad for us to not have one as
well.

The Contributor Covenant is probably the best generic option for a CoC
available to us. While Graphene does not have a "team", per se, writing
down the actual requirements for contributors is miles better than just
assuming good intentions.