Split headers from sources
While the GNOME best practices for libraries tend to mix sources with
header files, having the headers in a separate directory makes it easier
to distinguish public headers that will be installed from private headers
that will only be used internally. This is most useful when generating
headers that will be installed, like graphene-config.h and
graphene-version.h, as opposed to internal headers that won't, like
config.h.