MSVC builds: Disable SSE2 altogether for 32-bit x86 builds
It seems that __vectorcall/ABI is improved in Visual Studio 2017/2019
for 32-bit x86 builds (and tests) of Graphene itself, but is not
robust enough for more involved programs, such as GTK master, to run
without encountering strange access violations.  As a result, disable
SSE2 intrinsics for 32-bit x86 for now until the __vectorcall/ABI
situation gets improved on later Visual Studio iterations.

This affects builds up to the latest Visual Studio 2019 16.4.x,
unfortunately.