MSVC builds: Error out on C4819 if -utf-8 is not supported
The -utf-8 compiler flag is only supported since Visual Studio 2015, so
if that is not supported, fall back to the former behavior of erroring
out on C4819, so that we don't mis-compile things.

Note that this means we can build and run Graphene with Visual Studio
as-is on Visual Studio 2013 and later, but for building the tests,
Visual Studio 2015 or later is required, since Visual Studio 2015 and
later is required to build mutest.