meson.build: Warn about tests for pre-MSVC 2015 builds
We can actually build and use Graphene on Visual Studio 2013 and later,
but it does not support enough C99 for the dependent mutest test library
that is used for the tests.

Issue a warning that one may proceed building Graphene using Visual
Studio 2013, but it will come at a cost of having to disable building
and running the test programs.