meson: expose graphene_gobject_dep variable for graphene-gobject
Expose a separate variable for graphene-gobject for subproject usage
support, so the superproject and sibling subprojects can check for this
directly as part of the dependency() fallback logic instead of having to
poke at the build_gobject subproject variable. This mirrors the pkg-config
lookup logic where we have a dedicated .pc file for graphene-gobject as well.
If gobject support was disabled the variable will be set to a not-found dep.