Euler: Fix matrix orientation in euler matrix (de)composition
The euler_to_matrix and matrix_to_euler code is based on:

https://github.com/matthew-brett/transforms3d/blob/master/transforms3d/euler.py

This is num-py and has the same row-major matrixes like graphene. However
that code is based on column vectors being multiplied with the matrix on
the left:

> This module implements intrinsic and extrinsic axes, with standard conventions
> for axes `i`, `j`, `k`.  We assume that the matrix should be applied on the
> left of the vector, and right-handed coordinate systems.  To get the matrix to
> apply on the right of the vector, you need the transpose of the matrix we
> supply here, by the matrix transpose rule: $(M . V)^T = V^T M^T$.

To do the transpose we just change the matrix indexing macro (M) to swap
the row/column.