ARM NEON intrinsics: Add Visual Studio support
While we do use intrinsics for ARM NEON support, they are done in GCC-specific
ways that don't work with Visual Studio builds.

This adds such support that will work without the presence of GCCisms, but as
a result we need to use inline functions for them, since that's the best we can
do in this situation.