Add alignment macros
Different compilers have different placements for compiler attributes:
GCC-like uses postfix notation, while MSVC uses a prefix one.

We kind of coasted on GCC being gracious enough to support the same
placement as MSVC, but it'd be better if we used the postfix notation,
to ensure things work as intended.