ray: simplify nan checking by creating a macro.
Avoids the #ifdef forest and code duplication resulting from it.
There was mismatch in the two code paths, see [1] and [2], and this
commit avoids repeating the same mistake.

While there, use isnan() instead of fpclassify() == FP_NAN for the case
where isnanf() isn't available.

We use isnanf() (if available) instead of isnan() due to [3]:

    I initially used isinf() and isnan(), but those ended up breaking
    when using GCC because it tried to promote floats to doubles, and
    the results wouldn't match any more—especially when using GCC
    vectorisation.
    It could very well be a bug in GCC.

[1] https://github.com/ebassi/graphene/issues/223
[2] https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/3976
[3] https://github.com/ebassi/graphene/pull/174#issuecomment-561325167