meson.build: Update cflags for clang-cl
clang-cl follows some of the compiler flags that are used for GCC, but not all
them, so make sure that we don't try to use the compiler flags that it does not
understand.

Since the Visual Studio headers and libraries that clang-cl uses supported
SSE 4.1 for a very long while, even before the inception of clang-cl, always
pass in '-msse4.1' in the cflags when we are building with clang-cl (and it is
needed, otherwise the build fails).