graphene-simd4f.h: Simplify initializing SSE SIMD data arrays
Use the macros that we just added to initialize the graphene_simd4f_t arrays
with the floats that we pass into graphene_simd4f_init*() as applicable.  This
especially simplifies the code for Visual Studio since we already require C99
support for building and using Graphene, and we can reduce some code
duplication.