Use single vec4 for box2d
Pack the minimum and maximum vertices into a single vector, and use
merge low/high operators to extract the values. This reduces the size of
a box2d instance, and makes it more compact. It's also more efficient to
extract both vertices when passing them to GL, by using a single read.