box2d: Add inlined intersection operator
If we want to just know if two boxes intersect each other, we can do
that with an inlined function.

It's not the fastest possible function, because it still requires
calling into graphene_box2d_get_minmax(), as the calling code does not
have access to the private fields of the graphene_box2d_t type.